Deleting a product

+++ hello-market-server/controllers/productController.js
@@ -2,6 +2,7 @@
 const models = require('../models')
 const multer = require('multer')
 const path = require('path')
+const fs = require('fs');
 const { validationResult } = require('express-validator');
 
 // Configure multer for file storage
@@ -71,7 +72,23 @@ exports.deleteProduct = async (req, res) => {
   }
 
   const productId = req.params.productId
+
   try {
+
+    // find the product to get the image path 
+    const product = await models.Product.findByPk(productId) 
+    if(!product) {
+      return res.status(404).json({ message: 'Product not found', success: false });
+    }
+
+    const fileName = getFileNameFromUrl(product.photo_url)
+    console.log(fileName)
+
+    // delete the associated image 
+    //await deleteImageFile(fileName)
+
+    //return 
+
     const result = await models.Product.destroy({
       where: {
         id: productId
@@ -82,11 +99,56 @@ exports.deleteProduct = async (req, res) => {
       return res.status(404).json({ message: 'Product not found', success: false });
     }
 
+    // delete the associated image 
+    await deleteImageFile(fileName)
+
     return res.status(200).json({ message: `Product with ID ${productId} deleted successfully`, success: true });
+
   } catch (error) {
-    return res.status(500).json({ message: `Error deleting product ${error.message} ` });
+    return res.status(500).json({ message: `Error deleting product ${error.message} `, success: false });
   }
 }
+
+const getFileNameFromUrl = (photoUrl) => {
+  try {
+    const url = new URL(photoUrl); // Parse the URL
+    const fileName = path.basename(url.pathname); // Get the file name from the path
+    return fileName;
+  } catch (error) {
+    console.error('Invalid URL:', error);
+    return null;
+  }
+}
+
+// Helper function to delete image file
+const deleteImageFile = (fileName) => {
+  return new Promise((resolve, reject) => {
+    if (!fileName) {
+      return resolve(); // No image to delete
+    }
+
+    const fullImagePath = path.join(__dirname, '../uploads', fileName);
+
+    // Check if the file exists before trying to delete it
+    fs.access(fullImagePath, fs.constants.F_OK, (accessErr) => {
+      if (accessErr) {
+        console.error('File does not exist:', fullImagePath);
+        return resolve(); // If the file doesn't exist, we consider it successfully "deleted"
+      }
+
+      // If the file exists, delete it
+      fs.unlink(fullImagePath, (unlinkErr) => {
+        if (unlinkErr) {
+          console.error('Error deleting image:', unlinkErr);
+          return reject(unlinkErr);
+        } else {
+          console.log('Image deleted successfully');
+          return resolve();
+        }
+      });
+    });
+  });
+};
 
 exports.getMyProducts = async (req, res) => {
   try {
